order package added

+++ src/main/java/com/example/springbootbe/user/User.java
@@ -24,6 +24,6 @@ public class User {
 
     private String phone;
 
-    private Long userStatus;
+    private Integer userStatus;
 
 }
